use explicit write+exec mapping on Android
On Linux, by default libffi relies on malloc() memory to have
write+execute permissions, which is usually comes as a side effect of
making the stack executable (either by execstack linker option or
by compiling and linking some .S files; libffi linked programs get
linked to such objects).

On Android, at least on some versions of NDK for ARM, it is not enough
to just link some objects compiled from .S (or libffi) to enable
execstack and to get executable heap, so libffi just crashes in such
configuration. It can be argued it's not even desirable to make the
whole heap executable because of security implications, so enable
FFI_MMAP_EXEC_WRIT to make only needed memory executable instead.