Suppress error console output duing configure in case not an ELF file
On Cygwin, MSYS and MinGW readelf is available, but files generated are no
ELF files. This results in an Error output on console:
'checking whether .eh_frame section should be read-only... readelf: Error: Not an ELF file - it has the wrong magic bytes at the start
yes'
With this patch the error output is suppressed resulting in only
'checking whether .eh_frame section should be read-only... yes'
Setting of related variables is not changed.